ml-worker/user-vector/requirements.txt
# requirements.txt
pinecone-client==3.1.0
redis==5.0.1
pymongo==4.6.1
numpy==1.26.4